ATTENDANCE_WITH_FACE_DETECTION/requirements.txt
opencv-python
pillow
opencv-contrib-python
passlib
shutil
os
numpy
pickle
threading
openpyxl